finish merging / cleaning
