+++ src/Components/Navigation/MainNav.js
@@ -2,7 +2,7 @@ import * as React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import BottomNav from "./BottomNav";
 import CreateEvent from "../Screens/CreateEvent";
-import Messages from "../Screens/Messages";
+import Chat from "../Screens/Chat";
 
 const MainStack = createNativeStackNavigator();
 
@@ -11,7 +11,7 @@ function MainNav() {
     <MainStack.Navigator screenOptions={{ headerShown: false }}>
       <MainStack.Screen name="Root Nav" component={BottomNav} />
       <MainStack.Screen name="CreateEvent" component={CreateEvent} />
-      <MainStack.Screen name="Messages" component={Messages} options={{ headerShown: true }} />
+      <MainStack.Screen name="Chat" component={Chat} options={{ headerShown: true }} />
     </MainStack.Navigator>
   );
 }
