+++ src/Components/Screens/Profile.js
@@ -2,14 +2,7 @@ import { View, StyleSheet } from "react-native";
 import { useContext } from "react";
 import { UserContext } from "../Context/UserContext";
 import SignOut from "./SignOut";
-import {
-  Layout,
-  Card,
-  Text,
-  Button,
-  Avatar,
-  Divider,
-} from "@ui-kitten/components";
+import { Layout, Card, Text, Button, Avatar, Divider } from "@ui-kitten/components";
 import { Ionicons } from "@expo/vector-icons";
 import { Controller } from "react-hook-form";
 
@@ -21,11 +14,7 @@ function Profile() {
       <View style={styles.container}>
         <SignOut />
         <Card style={styles.profileContainer}>
-          <Avatar
-            size="giant"
-            source={require("../../../assets/avatars/Avatar3.png")}
-            style={styles.content}
-          />
+          <Avatar size="giant" source={require("../../../assets/avatars/Avatar3.png")} style={styles.content} />
           <Text category="h1" status="info">
             John Doe
           </Text>
@@ -35,12 +24,7 @@ function Profile() {
           </Text>
         </Card>
         <Card style={styles.contentContainer}>
-          <Ionicons
-            name="create"
-            style={styles.editIcon}
-            size={25}
-            onPress={console.log("edit favourites")}
-          />
+          <Ionicons name="create" style={styles.editIcon} size={25} onPress={console.log("edit favourites")} />
           <Text category="s2" style={styles.editDescription}>
             Edit
           </Text>
@@ -111,29 +95,28 @@ const styles = StyleSheet.create({
     borderTopWidth: 4,
     boxShadow: "0px 0px 5px 5px rgba(0,0,0,0.9)",
     position: "relative",
+    padding: 5,
   },
   editIcon: {
     top: 2,
     right: -60,
-    index: 1,
     position: "absolute",
+    color: "white",
   },
   editDescription: {
     top: 8,
     right: -35,
-    index: 1,
     position: "absolute",
   },
   createIcon: {
     top: 1,
     right: -75,
-    index: 1,
     position: "absolute",
+    color: "white",
   },
   createDescription: {
     top: 5,
     right: -50,
-    index: 1,
     position: "absolute",
   },
 });
